Create a BankAccount class with a balance
A Bank Account class has been created, which takes the account balance upon creation as an argument and stores it as an instance variable. Rather than hardcoding the balance as 0, this allows for possible future cases where new accounts are created with an existing balance- for example, if you are changing banks and transfering your funds.