Refactor TransactionHistory to own class
TransactionHistory has been refactored to it's own class. Within the class there is a list of transactions, and an add transaction method where the completed transaction is added to the history. The TransactionHistory is initialized upon creation of the BankAccount, and after each deposit or withdrawal the private method complete_transaction is called which inserts the transaction into history.